requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11.15
django-common-helpers==0.9.2
django-compat==1.0.15
django-cron==0.5.1
django-multiselectfield==0.1.8
humanize==0.5.1
idna==2.7
pytz==2018.5
requests==2.19.1
RPi.GPIO==0.6.3
six==1.11.0
urllib3==1.23
